backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
pymongo==4.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
googlemaps==4.10.0
haversine==2.8.0
ortools==9.7.2996
pytest==7.4.0
black==23.7.0
flake8==6.1.0
